requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
langsmith>=0.0.34
langchain-community>=0.0.10
python-dotenv>=1.0.0
docarray>=0.32.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.2
langchain-groq>=0.0.5 